+++ system_tests/system_test_harness.py
@@ -34,6 +34,7 @@ main_menubar = None
 
 tests_executed = 0
 test_failures = 0
+test_warnings = 0
 
 # ------------------------------------------------------------------------------
 # Global variables to track the 'one up' train idenntifier for every time
@@ -52,6 +53,11 @@ def raise_test_error(message):
     caller = getframeinfo(stack()[2][0])
     logging.error("Line %d of %s: %s" % (caller.lineno,basename(caller.filename),message))     
 
+def raise_test_warning(message):
+    global logging
+    caller = getframeinfo(stack()[2][0])
+    logging.warning("Line %d of %s: %s" % (caller.lineno,basename(caller.filename),message))     
+
 # ------------------------------------------------------------------------------
 # Function to initialise the test harness and load a schematic file (to test)
 # Note that if a filename is not specified then it will open a dialogue to
@@ -61,13 +67,12 @@ def raise_test_error(message):
 
 def initialise_test_harness(filename=None):
     global main_menubar, root
-    global test_failures, tests_executed
-    root = tkinter.Tk()
     logging.basicConfig(format='%(levelname)s: %(message)s')
     logging.getLogger().setLevel(logging.WARNING)
-    main_menubar = menubar.main_menubar(root)
-    schematic.create_canvas(root, main_menubar.handle_canvas_event)
-    main_menubar.initialise_editor()
+    if root is None:
+        root = tkinter.Tk()
+        main_menubar = menubar.main_menubar(root)
+        schematic.create_canvas(root, main_menubar.handle_canvas_event)
     print ("System Tests: Load File: '",filename,"'")
     main_menubar.load_schematic(filename)
     print ("System Tests: Setting Default Layout State")
@@ -83,7 +88,7 @@ def initialise_test_harness(filename=None):
 
 def complete_tests():
     print ("Tests Run:",tests_executed,"  Tests Passed:",
-              tests_executed-test_failures,"  Test failures", test_failures)
+              tests_executed-test_failures,"  Test failures",test_failures,"  Test Warnings",test_warnings)
     root.mainloop()
 
 # ------------------------------------------------------------------------------
@@ -102,81 +107,135 @@ def sleep(sleep_time):
 # ------------------------------------------------------------------------------
 
 def set_signals_on(*sigids):
+    global test_warnings
     for sigid in sigids:
-        if signals.signal_clear(sigid):
+        if str(sigid) not in signals_common.signals.keys():
+            raise_test_warning ("Signal:"+str(sigid)+" - does not exist on the schematic")
+            test_warnings = test_warnings+1
+        elif signals.signal_clear(sigid):
             signals_common.signal_button_event(sigid)
     root.update()
 
 def set_signals_off(*sigids):
+    global test_warnings
     for sigid in sigids:
-        if not signals.signal_clear(sigid):
+        if str(sigid) not in signals_common.signals.keys():
+            raise_test_warning ("Signal:"+str(sigid)+" - does not exist on the schematic")
+            test_warnings = test_warnings+1
+        elif not signals.signal_clear(sigid):
             signals_common.signal_button_event(sigid)
     root.update()
 
 def set_subsidaries_on(*sigids):
+    global test_warnings
     for sigid in sigids:
-        if not signals_common.signals[str(sigid)]["hassubsidary"]:
+        if str(sigid) not in signals_common.signals.keys():
+            raise_test_warning ("Signal:"+str(sigid)+" - does not exist on the schematic")
+            test_warnings = test_warnings+1
+        elif not signals_common.signals[str(sigid)]["hassubsidary"]:
             raise_test_error ("Signal:"+str(sigid)+" - Failed set_subsidaries_on (not supported)")
         elif signals.subsidary_clear(sigid):
             signals_common.subsidary_button_event(sigid)
     root.update()
 
 def set_subsidaries_off(*sigids):
+    global test_warnings
     for sigid in sigids:
-        if not signals_common.signals[str(sigid)]["hassubsidary"]:
+        if str(sigid) not in signals_common.signals.keys():
+            raise_test_warning ("Signal:"+str(sigid)+" - does not exist on the schematic")
+            test_warnings = test_warnings+1
+        elif not signals_common.signals[str(sigid)]["hassubsidary"]:
             raise_test_error ("Signal:"+str(sigid)+" - Failed set_subsidaries_on (not supported)")
         elif not signals.subsidary_clear(sigid):
             signals_common.subsidary_button_event(sigid)
     root.update()
 
+def trigger_signals_passed(*sigids):
+    global test_warnings
+    for sigid in sigids:
+        if str(sigid) not in signals_common.signals.keys():
+            raise_test_warning ("Signal:"+str(sigid)+" - does not exist on the schematic")
+            test_warnings = test_warnings+1
+        else:
+            signals_common.sig_passed_button_event(sigid)
+    root.update()
+
+def trigger_signals_released(*sigids):
+    global test_warnings
+    for sigid in sigids:
+        if str(sigid) not in signals_common.signals.keys():
+            raise_test_warning ("Signal:"+str(sigid)+" - does not exist on the schematic")
+            test_warnings = test_warnings+1
+        else:
+            signals_common.approach_release_button_event(sigid)
+    root.update()
+    
 def set_points_switched(*pointids):
+    global test_warnings
     for pointid in pointids:
-        if not points.point_switched(pointid):
+        if str(pointid) not in points.points.keys():
+            raise_test_warning ("Point:"+str(pointid)+" - does not exist on the schematic")
+            test_warnings = test_warnings+1
+        elif not points.point_switched(pointid):
             points.change_button_event(pointid)
     root.update()
 
 def set_points_normal(*pointids):
+    global test_warnings
     for pointid in pointids:
-        if points.point_switched(pointid):
+        if str(pointid) not in points.points.keys():
+            raise_test_warning ("Point:"+str(pointid)+" - does not exist on the schematic")
+            test_warnings = test_warnings+1
+        elif points.point_switched(pointid):
             points.change_button_event(pointid)
 
 def set_fpls_on(*pointids):
+    global test_warnings
     for pointid in pointids:
-        if not points.fpl_active(pointid):
+        if str(pointid) not in points.points.keys():
+            raise_test_warning ("Point:"+str(pointid)+" - does not exist on the schematic")
+            test_warnings = test_warnings+1
+        elif not points.fpl_active(pointid):
             points.fpl_button_event(pointid)
     root.update()
 
 def set_fpls_off(*pointids):
+    global test_warnings
     for pointid in pointids:
-        if points.fpl_active(pointid):
+        if str(pointid) not in points.points.keys():
+            raise_test_warning ("Point:"+str(pointid)+" - does not exist on the schematic")
+            test_warnings = test_warnings+1
+        elif points.fpl_active(pointid):
             points.fpl_button_event(pointid)
     root.update()
 
 def set_sections_occupied(*sectionids):
-    global train_identifier
-    for sectionid in sectionids:
-        if not track_sections.section_occupied(sectionid):
-            track_sections.clear_section_occupied(sectionid,str(train_identifier))
-            track_sections.section_button_event(sectionid)
+    global test_warnings, train_identifier
+    for secid in sectionids:
+        if str(secid) not in track_sections.sections.keys():
+            raise_test_warning ("Section:"+str(secid)+" - does not exist on the schematic")
+            test_warnings = test_warnings+1
         else:
-            track_sections.set_section_occupied(sectionid,str(train_identifier))
-        train_identifier=train_identifier+1
+            if not track_sections.section_occupied(secid):
+                track_sections.clear_section_occupied(secid,str(train_identifier))
+                track_sections.section_button_event(secid)
+            else:
+                track_sections.set_section_occupied(secid,str(train_identifier))
+            train_identifier=train_identifier+1
     root.update()
     
 def set_sections_clear(*sectionids):
-    for sectionid in sectionids:
-        if track_sections.section_occupied(sectionid):
-            track_sections.section_button_event(sectionid)
-    root.update()
-    
-def trigger_signals_passed(*sigids):
-    for sigid in sigids:
-        signals_common.sig_passed_button_event(sigid)
-    root.update()
-
-def trigger_signals_released(*sigids):
-    for sigid in sigids:
-        signals_common.approach_release_button_event(sigid)
+    global test_warnings, train_identifier
+    for secid in sectionids:
+        if str(secid) not in track_sections.sections.keys():
+            raise_test_warning ("Section:"+str(secid)+" - does not exist on the schematic")
+            test_warnings = test_warnings+1
+        else:
+            if track_sections.section_occupied(secid):
+                track_sections.section_button_event(secid)
+            else:
+                track_sections.clear_section_occupied(secid,str(train_identifier))
+            train_identifier=train_identifier+1
     root.update()
     
 # ------------------------------------------------------------------------------
@@ -184,81 +243,109 @@ def trigger_signals_released(*sigids):
 # ------------------------------------------------------------------------------
 
 def assert_points_locked(*pointids):
-    global test_failures, tests_executed
+    global test_warnings, test_failures, tests_executed
     for pointid in pointids:
-        if not points.points[str(pointid)]["locked"]:
+        if str(pointid) not in points.points.keys():
+            raise_test_warning ("Point:"+str(pointid)+" - does not exist on the schematic")
+            test_warnings = test_warnings+1
+        elif not points.points[str(pointid)]["locked"]:
             raise_test_error ("Point:"+str(pointid)+" - Failed assert_points_locked test")
             test_failures = test_failures+1
         tests_executed = tests_executed+1
     
 def assert_points_unlocked(*pointids):
-    global test_failures, tests_executed
+    global test_warnings, test_failures, tests_executed
     for pointid in pointids:
-        if points.points[str(pointid)]["locked"]:
+        if str(pointid) not in points.points.keys():
+            raise_test_warning ("Point:"+str(pointid)+" - does not exist on the schematic")
+            test_warnings = test_warnings+1
+        elif points.points[str(pointid)]["locked"]:
             raise_test_error ("Point:"+str(pointid)+" - Failed assert_points_unlocked test")
             test_failures = test_failures+1
         tests_executed = tests_executed+1
 
 def assert_signals_locked(*sigids):
-    global test_failures, tests_executed
+    global test_warnings, test_failures, tests_executed
     for sigid in sigids:
-        if not signals_common.signals[str(sigid)]["siglocked"]:
+        if str(sigid) not in signals_common.signals.keys():
+            raise_test_warning ("Signal:"+str(sigid)+" - does not exist on the schematic")
+            test_warnings = test_warnings+1
+        elif not signals_common.signals[str(sigid)]["siglocked"]:
             raise_test_error ("Signal:"+str(sigid)+" - Failed assert_signals_locked test")
             test_failures = test_failures+1
         tests_executed = tests_executed+1
             
 def assert_signals_unlocked(*sigids):
-    global test_failures, tests_executed
+    global test_warnings, test_failures, tests_executed
     for sigid in sigids:
-        if signals_common.signals[str(sigid)]["siglocked"]:
+        if str(sigid) not in signals_common.signals.keys():
+            raise_test_warning ("Signal:"+str(sigid)+" - does not exist on the schematic")
+            test_warnings = test_warnings+1
+        elif signals_common.signals[str(sigid)]["siglocked"]:
             raise_test_error ("Signal:"+str(sigid)+" - Failed assert_signals_unlocked test")
             test_failures = test_failures+1
         tests_executed = tests_executed+1
 
 def assert_subsidaries_locked(*sigids):
-    global test_failures, tests_executed
+    global test_warnings, test_failures, tests_executed
     for sigid in sigids:
-        if not signals_common.signals[str(sigid)]["hassubsidary"]:
-            raise_test_error ("Signal:"+str(sigid)+" - Failed assert_subsidaries_locked test (not supported)")
+        if str(sigid) not in signals_common.signals.keys():
+            raise_test_warning ("Signal:"+str(sigid)+" - does not exist on the schematic")
+            test_warnings = test_warnings+1
+        elif not signals_common.signals[str(sigid)]["hassubsidary"]:
+            raise_test_warning ("Signal:"+str(sigid)+" - Failed assert_subsidaries_locked test (not supported)")
+            test_warnings = test_warnings+1
         elif not signals_common.signals[str(sigid)]["sublocked"]:
             raise_test_error ("Signal:"+str(sigid)+" - Failed assert_subsidaries_locked test")
             test_failures = test_failures+1
         tests_executed = tests_executed+1
             
 def assert_subsidaries_unlocked(*sigids):
-    global test_failures, tests_executed
+    global test_warnings, test_failures, tests_executed
     for sigid in sigids:
-        if not signals_common.signals[str(sigid)]["hassubsidary"]:
-            raise_test_error ("Signal:"+str(sigid)+" - Failed assert_subsidaries_unlocked test (not supported)")
-        if signals_common.signals[str(sigid)]["sublocked"]:
+        if str(sigid) not in signals_common.signals.keys():
+            raise_test_warning ("Signal:"+str(sigid)+" - does not exist on the schematic")
+            test_warnings = test_warnings+1
+        elif not signals_common.signals[str(sigid)]["hassubsidary"]:
+            raise_test_warning ("Signal:"+str(sigid)+" - Failed assert_subsidaries_unlocked test (not supported)")
+            test_warnings = test_warnings+1
+        elif signals_common.signals[str(sigid)]["sublocked"]:
             raise_test_error ("Signal:"+str(sigid)+" - Failed assert_subsidaries_unlocked test")
             test_failures = test_failures+1
         tests_executed = tests_executed+1
         
 def assert_signals_override_set(*sigids):
-    global test_failures, tests_executed
+    global test_warnings, test_failures, tests_executed
     for sigid in sigids:
-        if not signals_common.signals[str(sigid)]["override"]:
+        if str(sigid) not in signals_common.signals.keys():
+            raise_test_warning ("Signal:"+str(sigid)+" - does not exist on the schematic")
+            test_warnings = test_warnings+1
+        elif not signals_common.signals[str(sigid)]["override"]:
             raise_test_error ("Signal:"+str(sigid)+" - Failed assert_signals_override_set test")
             test_failures = test_failures+1
         tests_executed = tests_executed+1
         
 def assert_signals_override_clear(*sigids):
-    global test_failures, tests_executed
+    global test_warnings, test_failures, tests_executed
     for sigid in sigids:
-        if signals_common.signals[str(sigid)]["override"]:
+        if str(sigid) not in signals_common.signals.keys():
+            raise_test_warning ("Signal:"+str(sigid)+" - does not exist on the schematic")
+            test_warnings = test_warnings+1
+        elif signals_common.signals[str(sigid)]["override"]:
             raise_test_error ("Signal:"+str(sigid)+" - Failed assert_signals_override_clear test")
             test_failures = test_failures+1
         tests_executed = tests_executed+1
 
 def assert_signals_app_cntl_set(*sigids):
-    global test_failures, tests_executed
+    global test_warnings, test_failures, tests_executed
     for sigid in sigids:
-        # Check that the signal supports approach control to prevent an exception
-        if ( "releaseonred" not in signals_common.signals[str(sigid)].keys() or
+        if str(sigid) not in signals_common.signals.keys():
+            raise_test_warning ("Signal:"+str(sigid)+" - does not exist on the schematic")
+            test_warnings = test_warnings+1
+        elif ( "releaseonred" not in signals_common.signals[str(sigid)].keys() or
              "releaseonyel" not in signals_common.signals[str(sigid)].keys() ):
-            raise_test_error ("Signal:"+str(sigid)+" - Failed assert_signals_app_cntl_set test (not supported)")
-            test_failures = test_failures+1
+            raise_test_warning ("Signal:"+str(sigid)+" - Failed assert_signals_app_cntl_set test (not supported)")
+            test_warnings = test_warnings+1
         elif ( not signals_common.signals[str(sigid)]["releaseonred"] and
                not signals_common.signals[str(sigid)]["releaseonyel"] ):
             raise_test_error ("Signal:"+str(sigid)+" - Failed assert_signals_app_cntl_set test")
@@ -266,13 +353,15 @@ def assert_signals_app_cntl_set(*sigids):
         tests_executed = tests_executed+1
         
 def assert_signals_app_cntl_clear(*sigids):
-    global test_failures, tests_executed
+    global test_warnings, test_failures, tests_executed
     for sigid in sigids:
-        # Check that the signal supports approach control to prevent an exception
-        if ( "releaseonred" not in signals_common.signals[str(sigid)].keys() or
+        if str(sigid) not in signals_common.signals.keys():
+            raise_test_warning ("Signal:"+str(sigid)+" - does not exist on the schematic")
+            test_warnings = test_warnings+1
+        elif ( "releaseonred" not in signals_common.signals[str(sigid)].keys() or
              "releaseonyel" not in signals_common.signals[str(sigid)].keys() ):
-            raise_test_error ("Signal:"+str(sigid)+" - Failed assert_signals_app_cntl_clear test (not supported)")
-            test_failures = test_failures+1
+            raise_test_warning ("Signal:"+str(sigid)+" - Failed assert_signals_app_cntl_clear test (not supported)")
+            test_warnings = test_warnings+1
         elif ( signals_common.signals[str(sigid)]["releaseonred"] or
                signals_common.signals[str(sigid)]["releaseonyel"] ):
             raise_test_error ("Signal:"+str(sigid)+" - Failed assert_signals_app_cntl_clear test")
@@ -280,113 +369,156 @@ def assert_signals_app_cntl_clear(*sigids):
         tests_executed = tests_executed+1
 
 def assert_signals_DANGER(*sigids):
-    global test_failures, tests_executed
+    global test_warnings, test_failures, tests_executed
     for sigid in sigids:
-        if signals_common.signals[str(sigid)]["sigstate"]!=signals_common.signal_state_type.DANGER:
+        if str(sigid) not in signals_common.signals.keys():
+            raise_test_warning ("Signal:"+str(sigid)+" - does not exist on the schematic")
+            test_warnings = test_warnings+1
+        elif signals_common.signals[str(sigid)]["sigstate"]!=signals_common.signal_state_type.DANGER:
             raise_test_error ("Signal:"+str(sigid)+" - Failed assert_signals_DANGER test")
             test_failures = test_failures+1
         tests_executed = tests_executed+1
             
 def assert_signals_PROCEED(*sigids):
-    global test_failures, tests_executed
+    global test_warnings, test_failures, tests_executed
     for sigid in sigids:
-        if signals_common.signals[str(sigid)]["sigstate"]!=signals_common.signal_state_type.PROCEED:
+        if str(sigid) not in signals_common.signals.keys():
+            raise_test_warning ("Signal:"+str(sigid)+" - does not exist on the schematic")
+            test_warnings = test_warnings+1
+        elif signals_common.signals[str(sigid)]["sigstate"]!=signals_common.signal_state_type.PROCEED:
             raise_test_error ("Signal:"+str(sigid)+" - Failed assert_signals_PROCEED test")
             test_failures = test_failures+1
         tests_executed = tests_executed+1
             
 def assert_signals_CAUTION(*sigids):
-    global test_failures, tests_executed
+    global test_warnings, test_failures, tests_executed
     for sigid in sigids:
-        if signals_common.signals[str(sigid)]["sigstate"]!=signals_common.signal_state_type.CAUTION:
+        if str(sigid) not in signals_common.signals.keys():
+            raise_test_warning ("Signal:"+str(sigid)+" - does not exist on the schematic")
+            test_warnings = test_warnings+1
+        elif signals_common.signals[str(sigid)]["sigstate"]!=signals_common.signal_state_type.CAUTION:
             raise_test_error ("Signal:"+str(sigid)+" - Failed assert_signals_CAUTION test")
             test_failures = test_failures+1
         tests_executed = tests_executed+1
 
 def assert_signals_CAUTION_APP_CNTL(*sigids):
-    global test_failures, tests_executed
+    global test_warnings, test_failures, tests_executed
     for sigid in sigids:
-        if signals_common.signals[str(sigid)]["sigstate"]!=signals_common.signal_state_type.CAUTION_APP_CNTL:
+        if str(sigid) not in signals_common.signals.keys():
+            raise_test_warning ("Signal:"+str(sigid)+" - does not exist on the schematic")
+            test_warnings = test_warnings+1
+        elif signals_common.signals[str(sigid)]["sigstate"]!=signals_common.signal_state_type.CAUTION_APP_CNTL:
             raise_test_error ("Signal:"+str(sigid)+" - Failed assert_signals_CAUTION_APP_CNTL test")
             test_failures = test_failures+1
         tests_executed = tests_executed+1
 
 def assert_signals_PRELIM_CAUTION(*sigids):
-    global test_failures, tests_executed
+    global test_warnings, test_failures, tests_executed
     for sigid in sigids:
-        if signals_common.signals[str(sigid)]["sigstate"]!=signals_common.signal_state_type.PRELIM_CAUTION:
+        if str(sigid) not in signals_common.signals.keys():
+            raise_test_warning ("Signal:"+str(sigid)+" - does not exist on the schematic")
+            test_warnings = test_warnings+1
+        elif signals_common.signals[str(sigid)]["sigstate"]!=signals_common.signal_state_type.PRELIM_CAUTION:
             raise_test_error ("Signal:"+str(sigid)+" - Failed assert_signals_PRELIM_CAUTION test")
             test_failures = test_failures+1
         tests_executed = tests_executed+1
 
 def assert_signals_FLASH_CAUTION(*sigids):
-    global test_failures, tests_executed
+    global test_warnings, test_failures, tests_executed
     for sigid in sigids:
-        if signals_common.signals[str(sigid)]["sigstate"]!=signals_common.signal_state_type.FLASH_CAUTION:
+        if str(sigid) not in signals_common.signals.keys():
+            raise_test_warning ("Signal:"+str(sigid)+" - does not exist on the schematic")
+            test_warnings = test_warnings+1
+        elif signals_common.signals[str(sigid)]["sigstate"]!=signals_common.signal_state_type.FLASH_CAUTION:
             raise_test_error ("Signal:"+str(sigid)+" - Failed assert_signals_FLASH_CAUTION test")
             test_failures = test_failures+1
         tests_executed = tests_executed+1
 
 def assert_signals_FLASH_PRELIM_CAUTION(*sigids):
-    global test_failures, tests_executed
+    global test_warnings, test_failures, tests_executed
     for sigid in sigids:
-        if signals_common.signals[str(sigid)]["sigstate"]!=signals_common.signal_state_type.FLASH_PRELIM_CAUTION:
+        if str(sigid) not in signals_common.signals.keys():
+            raise_test_warning ("Signal:"+str(sigid)+" - does not exist on the schematic")
+            test_warnings = test_warnings+1
+        elif signals_common.signals[str(sigid)]["sigstate"]!=signals_common.signal_state_type.FLASH_PRELIM_CAUTION:
             raise_test_error ("Signal:"+str(sigid)+" - Failed assert_signals_FLASH_PRELIM_CAUTION test")
             test_failures = test_failures+1
         tests_executed = tests_executed+1
         
 def assert_signals_route_MAIN(*sigids):
-    global test_failures, tests_executed
+    global test_warnings, test_failures, tests_executed
     for sigid in sigids:
-        if signals_common.signals[str(sigid)]["routeset"]!=signals_common.route_type.MAIN:
+        if str(sigid) not in signals_common.signals.keys():
+            raise_test_warning ("Signal:"+str(sigid)+" - does not exist on the schematic")
+            test_warnings = test_warnings+1
+        elif signals_common.signals[str(sigid)]["routeset"]!=signals_common.route_type.MAIN:
             raise_test_error ("Signal:"+str(sigid)+" - Failed assert_signals_route_MAIN test")
             test_failures = test_failures+1
+        
         tests_executed = tests_executed+1
 
 def assert_signals_route_LH1(*sigids):
-    global test_failures, tests_executed
+    global test_warnings, test_failures, tests_executed
     for sigid in sigids:
-        if signals_common.signals[str(sigid)]["routeset"]!=signals_common.route_type.LH1:
+        if str(sigid) not in signals_common.signals.keys():
+            raise_test_warning ("Signal:"+str(sigid)+" - does not exist on the schematic")
+            test_warnings = test_warnings+1
+        elif signals_common.signals[str(sigid)]["routeset"]!=signals_common.route_type.LH1:
             raise_test_error ("Signal:"+str(sigid)+" - Failed assert_signals_route_LH1 test")
             test_failures = test_failures+1
         tests_executed = tests_executed+1
 
 def assert_signals_route_LH2(*sigids):
-    global test_failures, tests_executed
+    global test_warnings, test_failures, tests_executed
     for sigid in sigids:
-        if signals_common.signals[str(sigid)]["routeset"]!=signals_common.route_type.LH2:
+        if str(sigid) not in signals_common.signals.keys():
+            raise_test_warning ("Signal:"+str(sigid)+" - does not exist on the schematic")
+            test_warnings = test_warnings+1
+        elif signals_common.signals[str(sigid)]["routeset"]!=signals_common.route_type.LH2:
             raise_test_error ("Signal:"+str(sigid)+" - Failed assert_signals_route_LH2 test")
             test_failures = test_failures+1
         tests_executed = tests_executed+1
 
 def assert_signals_route_RH1(*sigids):
-    global test_failures, tests_executed
+    global test_warnings, test_failures, tests_executed
     for sigid in sigids:
-        if signals_common.signals[str(sigid)]["routeset"]!=signals_common.route_type.RH1:
+        if str(sigid) not in signals_common.signals.keys():
+            raise_test_warning ("Signal:"+str(sigid)+" - does not exist on the schematic")
+            test_warnings = test_warnings+1
+        elif signals_common.signals[str(sigid)]["routeset"]!=signals_common.route_type.RH1:
             raise_test_error ("Signal:"+str(sigid)+" - Failed assert_signals_route_RH1 test")
             test_failures = test_failures+1
         tests_executed = tests_executed+1
 
 def assert_signals_route_RH2(*sigids):
-    global test_failures, tests_executed
+    global test_warnings, test_failures, tests_executed
     for sigid in sigids:
-        if signals_common.signals[str(sigid)]["routeset"]!=signals_common.route_type.RH2:
+        if str(sigid) not in signals_common.signals.keys():
+            raise_test_warning ("Signal:"+str(sigid)+" - does not exist on the schematic")
+            test_warnings = test_warnings+1
+        elif signals_common.signals[str(sigid)]["routeset"]!=signals_common.route_type.RH2:
             raise_test_error ("Signal:"+str(sigid)+" - Failed assert_signals_route_RH2 test")
             test_failures = test_failures+1
         tests_executed = tests_executed+1
         
 def assert_sections_occupied(*secids):
-    global test_failures, tests_executed
+    global test_warnings, test_failures, tests_executed
     for secid in secids:
-        if not track_sections.sections[str(secid)]["occupied"]:
+        if str(secid) not in track_sections.sections.keys():
+            raise_test_warning ("Section:"+str(secid)+" - does not exist on the schematic")
+            test_warnings = test_warnings+1
+        elif not track_sections.sections[str(secid)]["occupied"]:
             raise_test_error ("Section:"+str(secid)+" - Failed assert_sections_occupied test")
             test_failures = test_failures+1
         tests_executed = tests_executed+1
 
 def assert_sections_clear(*secids):
-    global test_failures, tests_executed
+    global test_warnings, test_failures, tests_executed
     for secid in secids:
-        if track_sections.sections[str(secid)]["occupied"]:
+        if str(secid) not in track_sections.sections.keys():
+            raise_test_warning ("Section:"+str(secid)+" - does not exist on the schematic")
+            test_warnings = test_warnings+1
+        elif track_sections.sections[str(secid)]["occupied"]:
             raise_test_error ("Section:"+str(secid)+" - Failed assert_sections_clear test")
             test_failures = test_failures+1
         tests_executed = tests_executed+1
